requirements.txt
mcp-python
openai
requests
python-dotenv
